+++ sql/wsrep_applier.cc
@@ -269,8 +269,8 @@ wsrep_cb_status_t wsrep_apply_cb(void* const             ctx,
   TABLE *tmp;
   while ((tmp = thd->temporary_tables))
   {
-    WSREP_DEBUG("Applier %lu, has temporary tables: %s.%s",
-		thd->thread_id,
+    WSREP_DEBUG("Applier %lld, has temporary tables: %s.%s",
+		(longlong) thd->thread_id,
 		(tmp->s) ? tmp->s->db.str : "void",
 		(tmp->s) ? tmp->s->table_name.str : "void");
     close_temporary_table(thd, tmp, 1, 1);
